Do not start tests if $DFLOWFMDIR is not set.

+++ model_dflowfm_blackbox/tests/run_all_tests.sh
@@ -1,5 +1,9 @@
 #! /bin/sh
-set -x 
+if [ -z $DFLOWFMDIR ]; then
+    echo ERROR: No environment variable \$DFLOWFMDIR available.
+    echo use \"export DFLOWFMDIR={installation directory of D-Flow FM}\" before starting this script.
+    exit 1
+fi
 
 ./clean.sh
 rm -f *.log
